LPS-116566 Moves portlet keys to reuse it
