+++ modules/apps/style-book/style-book-web/src/main/java/com/liferay/style/book/web/internal/application/list/StyleBookPanelApp.java
@@ -18,7 +18,7 @@ import com.liferay.application.list.BasePanelApp;
 import com.liferay.application.list.PanelApp;
 import com.liferay.application.list.constants.PanelCategoryKeys;
 import com.liferay.portal.kernel.model.Portlet;
-import com.liferay.style.book.web.internal.constants.StyleBookPortletKeys;
+import com.liferay.style.book.constants.StyleBookPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
