+++ modules/apps/style-book/style-book-web/src/main/java/com/liferay/style/book/web/internal/portlet/StyleBookPortlet.java
@@ -16,7 +16,7 @@ package com.liferay.style.book.web.internal.portlet;
 
 import com.liferay.item.selector.ItemSelector;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.style.book.web.internal.constants.StyleBookPortletKeys;
+import com.liferay.style.book.constants.StyleBookPortletKeys;
 import com.liferay.style.book.web.internal.constants.StyleBookWebKeys;
 
 import java.io.IOException;
