+++ modules/apps/style-book/style-book-web/src/main/java/com/liferay/style/book/web/internal/portlet/action/UploadStyleBookEntryPreviewMVCActionCommand.java
@@ -17,7 +17,7 @@ package com.liferay.style.book.web.internal.portlet.action;
 import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.style.book.web.internal.constants.StyleBookPortletKeys;
+import com.liferay.style.book.constants.StyleBookPortletKeys;
 import com.liferay.style.book.web.internal.upload.StyleBookEntryPreviewUploadFileEntryHandler;
 import com.liferay.upload.UploadHandler;
